Imported 2.2.0-dfsg

+++ src/VBox/Devices/Storage/DrvHostBase.h
@@ -1,4 +1,4 @@
-/* $Id: DrvHostBase.h $ */
+/* $Id: DrvHostBase.h 18436 2009-03-28 02:32:20Z vboxsync $ */
 /** @file
  * DrvHostBase - Host base drive access driver.
  */
@@ -177,7 +177,7 @@ void DRVHostBaseMediaNotPresent(PDRVHOSTBASE pThis);
 DECLCALLBACK(void) DRVHostBaseDestruct(PPDMDRVINS pDrvIns);
 #ifdef RT_OS_DARWIN
 DECLCALLBACK(int) DRVHostBaseScsiCmd(PDRVHOSTBASE pThis, const uint8_t *pbCmd, size_t cbCmd, PDMBLOCKTXDIR enmTxDir,
-                                     void *pvBuf, size_t *pcbBuf, uint8_t *pbSense, size_t cbSense, uint32_t cTimeoutMillies);
+                                     void *pvBuf, uint32_t *pcbBuf, uint8_t *pbSense, size_t cbSense, uint32_t cTimeoutMillies);
 #endif
